Trying conftest option
